Moved app initialization to its own file to be imported from directly instead of __init__.py
