+++ manage.py
@@ -2,7 +2,7 @@
 
 from flask_script import Manager, Server
 
-from openfecwebapp import app
+from openfecwebapp.app import app
 
 manager = Manager(app)
 manager.add_command(
