+++ openfecwebapp/app.py
@@ -1,0 +1,3 @@
+from flask import Flask
+
+app = Flask('openfecwebapp', static_path='/static', static_folder='../dist')
